@@ -1,10 +1,9 @@
-name: better_player_example
-description: Better Player example usage app.
+name: example
+description: "Example to showcase use cases of Better Player."
+publish_to: 'none'
 version: 1.0.0+1
-publish_to: none
-
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.4.1 <4.0.0'
 
 dependencies:
   better_player:
@@ -13,9 +12,9 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  path_provider: ^2.0.10
-  visibility_detector: ^0.3.3
-  collection: ^1.16.0
+  path_provider: ^2.1.3
+  visibility_detector: ^0.4.0+2
+  collection: ^1.18.0
 
 dev_dependencies:
   flutter_test:
