@@ -1,30 +1,27 @@
 name: better_player
-description: Advanced video player based on video_player and Chewie. It's solves many typical use cases and it's easy to run.
-version: 0.0.83
-# Disabled because of warning from analyzer
-# authors:
-#  - Jakub Homlala <jhomlala@gmail.com>
+description: Advanced video player. It solves many typical use cases and it's easy to run.
+version: 0.0.84
 homepage: https://github.com/jhomlala/betterplayer
 documentation: https://jhomlala.github.io/betterplayer/
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=2.2.3"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.3.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.5
-  wakelock: ^0.6.2
-  meta: ^1.8.0
-  flutter_widget_from_html_core: ^0.8.5+3
-  visibility_detector: ^0.3.3
-  path_provider: ^2.0.11
-  collection: ^1.16.0
-  xml: ^6.1.0
+  cupertino_icons: ^1.0.8
+  wakelock_plus: ^1.2.7
+  meta: ^1.12.0
+  flutter_widget_from_html_core: ^0.15.1
+  visibility_detector: ^0.4.0+2
+  path_provider: ^2.1.3
+  collection: ^1.18.0
+  xml: ^6.5.0
 
 dev_dependencies:
-  lint: ^1.10.0
+  lint: ^2.3.0
   flutter_test:
     sdk: flutter
   flutter_localizations:
