@@ -1,5 +1,5 @@
 import 'package:better_player/better_player.dart';
-import 'package:better_player_example/constants.dart';
+import 'package:example/constants.dart';
 import 'package:flutter/material.dart';
 
 class DrmPage extends StatefulWidget {
